Add websocket connection for live refresh on local server
The local server (run by bin/serve) enables working incrementally on local
files referred by a codelab.
Everytime a codelab is changed, the modification is picked, corresponding
codelab is rebuilt and a signal is sent to all connected websockets to refresh
the website for this or those codelabs. This refresh is filtered client-side
to match corresponding tutorial name in the url. If not, it's a no-op.
We do reconnect the socket if it was able to connect successfully once and
then was disconnected. The effect is that on production server which won't have
that feature, we attempt to connect once, fail, but don't retry.