Kubeflow tutorial; improve intro, text wrapping
Condenses the introduction, moving the information about
each of the technologies used to a new section.

The Kubeflow tutorial wraps paragraphs poorly when published
to the web. This commit folds multiple sentences of the same
paragraph onto the same line. This allows the browser to re-
flow the paragraph as it needs to.